Refactored solver to be independenent of the solver page
