+++ client/src/components/Solver.css
