+++ client/src/pages/DemoSolverPage.jsx
@@ -1,0 +1,6 @@
+
+function DemoSolverPage({ }) {
+    return <></>
+}
+
+export default DemoSolverPage;
